change japanese code, jp into ja
