+++ test/testdata/ja.js
